Add `valid-bucket-name?` predicate.
Determines whether a name is a strictly valid name
(such as one recommended in the US, and required
elsewhere). By supplying #f for the optional argument
`dns-compliant?`, will say if the name would be valid
in the looser sense permitted for US buckets.

Although this isn't part of the fix for issue #16, it is
related.