Run ./make-doc.sh, update `manual.md` w/changes from commit 93642f3f
(Someday Racket's new package manager may add an equivalent to the doc
hosting provided by the old Planet. In the meantime, I've been using
GitHub's ability to host markdown files, and Scribble's ability to
generate them.

Which would be clearer if I'd bothered to provide a proper makefile
for this project.)