Add `test-data-exists?` and .travis.yml.
Use this to check if the test data file exists. If it does not, don't
even attempt to run tests that require the data.

Immediate motivation is to fix DrDr build issues. Also, this allows us
to use Travis CI (although obviously not to run the full tests that
interact with AWS).