s3: WIP for sig v4 issue #36
I need to do more testing. Also, I'm concerned about speed, due to the
need to sign the payload.

For now I think it will be acceptable that plain `put` signs the entire
payload up-front (not to use the chunked transfer upload), because
multipart-put is available and people ought to use that for larger
payloads, anyway (because it's faster due to parallel uploading and it's
recoverable). That is already "chunked" in the sense that there are
multiple parts.