Retry a few odd 200 Not Really OK responses
Amazon documents that S3 can, for delete-multiple, copy, and
complete-multipart-upload, return a 200 response code, but we must check
the response body to see if it actually succeeded. Although we were
already checking this, we ought to include these errors in what we
retry.

Rather than try to shoehorn this 200 Not OK notion into the
call/xxx-request/retry functions, handle it locally in each of the 3
spots. (This means there are "nested retries", but I believe this is
OK -- that is, Really OK.)